Move PeerTrust tests into pre-release test files.
* These tests can't work in released S.P.SM version 4.1.0 as needed product code has only been added since then, moving them into a new pre-release test file.
