+++ src/System.Private.ServiceModel/tests/Scenarios/Security/TransportSecurity/Https/HttpsTests.4.1.0.cs
@@ -4,18 +4,15 @@
 
 
 using System;
-using System.IdentityModel.Selectors;
-using System.Net;
 using System.Security.Cryptography.X509Certificates;
 using System.ServiceModel;
 using System.ServiceModel.Channels;
 using System.ServiceModel.Security;
 using System.Text;
-using System.Threading;
 using Infrastructure.Common;
 using Xunit;
 
-public class HttpsTests : ConditionalWcfTest
+public partial class HttpsTests : ConditionalWcfTest
 {
     // Client: CustomBinding set MessageVersion to Soap11
     // Server: BasicHttpsBinding default value is Soap11
@@ -242,7 +239,7 @@ public class HttpsTests : ConditionalWcfTest
     [Fact]
     [ActiveIssue(959)] // Server certificate validation not supported in NET Native
 #else
-    [ConditionalFact(nameof(Root_Certificate_Installed), 
+    [ConditionalFact(nameof(Root_Certificate_Installed),
                      nameof(Client_Certificate_Installed),
                      nameof(SSL_Available))]
 #endif
@@ -309,8 +306,8 @@ public class HttpsTests : ConditionalWcfTest
 #if FULLXUNIT_NOTSUPPORTED
     [Fact]
 #else
-    [ConditionalFact(nameof(Root_Certificate_Installed), 
-                     nameof(Client_Certificate_Installed), 
+    [ConditionalFact(nameof(Root_Certificate_Installed),
+                     nameof(Client_Certificate_Installed),
                      nameof(Server_Accepts_Certificates),
                      nameof(SSL_Available))]
 #endif
@@ -358,137 +355,6 @@ public class HttpsTests : ConditionalWcfTest
                 StoreName.My,
                 X509FindType.FindByThumbprint,
                 clientCertThumb);
-
-            serviceProxy = factory.CreateChannel();
-
-            // *** EXECUTE *** \\
-            string result = serviceProxy.Echo(testString);
-
-            // *** VALIDATE *** \\
-            Assert.Equal(testString, result);
-
-            // *** CLEANUP *** \\
-            ((ICommunicationObject)serviceProxy).Close();
-            factory.Close();
-        }
-        finally
-        {
-            // *** ENSURE CLEANUP *** \\
-            ScenarioTestHelpers.CloseCommunicationObjects((ICommunicationObject)serviceProxy, factory);
-        }
-    }
-
-#if FULLXUNIT_NOTSUPPORTED
-    [Fact]
-    [ActiveIssue(959)] // Server certificate validation not supported in NET Native
-#else
-    [ConditionalFact(nameof(Root_Certificate_Installed), 
-                     nameof(Client_Certificate_Installed),
-                     nameof(Peer_Certificate_Installed),
-                     nameof(SSL_Available))]
-#endif
-    [OuterLoop]
-    public static void Https_SecModeTrans_ClientCredTypeNone_ServerCertValModePeerTrust_EchoString()
-    {
-#if FULLXUNIT_NOTSUPPORTED
-        bool root_Certificate_Installed = Root_Certificate_Installed();
-        bool client_Certificate_Installed = Client_Certificate_Installed();
-        bool peer_Certificate_Installed = Peer_Certificate_Installed();
-        bool ssl_Available = SSL_Available();
-
-        if (!root_Certificate_Installed || 
-            !client_Certificate_Installed ||
-            !peer_Certificate_Installed ||
-            !ssl_Available)
-        {
-            Console.WriteLine("---- Test SKIPPED --------------");
-            Console.WriteLine("Attempting to run the test in ToF, a ConditionalFact evaluated as FALSE.");
-            Console.WriteLine("Root_Certificate_Installed evaluated as {0}", root_Certificate_Installed);
-            Console.WriteLine("Client_Certificate_Installed evaluated as {0}", client_Certificate_Installed);
-            Console.WriteLine("Peer_Certificate_Installed evaluated as {0}", peer_Certificate_Installed);
-            Console.WriteLine("SSL_Available evaluated as {0}", ssl_Available);
-            return;
-        }
-#endif
-        EndpointAddress endpointAddress = null;
-        string testString = "Hello";
-        ChannelFactory<IWcfService> factory = null;
-        IWcfService serviceProxy = null;
-
-        try
-        {
-            // *** SETUP *** \\
-            BasicHttpsBinding binding = new BasicHttpsBinding(BasicHttpsSecurityMode.Transport);
-            binding.Security.Transport.ClientCredentialType = HttpClientCredentialType.None;
-            endpointAddress = new EndpointAddress(new Uri(Endpoints.Https_SecModeTrans_ClientCredTypeNone_ServerCertValModePeerTrust_Address));
-
-            factory = new ChannelFactory<IWcfService>(binding, endpointAddress);
-            factory.Credentials.ServiceCertificate.SslCertificateAuthentication = new X509ServiceCertificateAuthentication();
-            factory.Credentials.ServiceCertificate.SslCertificateAuthentication.CertificateValidationMode = X509CertificateValidationMode.PeerTrust;
-
-            serviceProxy = factory.CreateChannel();
-
-            // *** EXECUTE *** \\
-            string result = serviceProxy.Echo(testString);
-
-            // *** VALIDATE *** \\
-            Assert.Equal(testString, result);
-
-            // *** CLEANUP *** \\
-            ((ICommunicationObject)serviceProxy).Close();
-            factory.Close();
-        }
-        finally
-        {
-            // *** ENSURE CLEANUP *** \\
-            ScenarioTestHelpers.CloseCommunicationObjects((ICommunicationObject)serviceProxy, factory);
-        }
-    }
-
-    // This test is expected to validate the service certificate using PeerTrust with X509CertificateValidationMode set to PeerOrChainTrust.
-#if FULLXUNIT_NOTSUPPORTED
-    [Fact]
-    [ActiveIssue(959)] // Server certificate validation not supported in NET Native
-#else
-    [ConditionalFact(nameof(Root_Certificate_Installed), 
-                     nameof(Client_Certificate_Installed),
-                     nameof(SSL_Available))]
-#endif
-    [OuterLoop]
-    public static void Https_SecModeTrans_ClientCredTypeNone_ServerCertValModePeerOrChainTrust_EchoString()
-    {
-#if FULLXUNIT_NOTSUPPORTED
-        bool root_Certificate_Installed = Root_Certificate_Installed();
-        bool client_Certificate_Installed = Client_Certificate_Installed();
-        bool ssl_Available = SSL_Available();
-
-        if (!root_Certificate_Installed || 
-            !client_Certificate_Installed ||
-            !ssl_Available)
-        {
-            Console.WriteLine("---- Test SKIPPED --------------");
-            Console.WriteLine("Attempting to run the test in ToF, a ConditionalFact evaluated as FALSE.");
-            Console.WriteLine("Root_Certificate_Installed evaluated as {0}", root_Certificate_Installed);
-            Console.WriteLine("Client_Certificate_Installed evaluated as {0}", client_Certificate_Installed);
-            Console.WriteLine("SSL_Available evaluated as {0}", ssl_Available);
-            return;
-        }
-#endif
-        EndpointAddress endpointAddress = null;
-        string testString = "Hello";
-        ChannelFactory<IWcfService> factory = null;
-        IWcfService serviceProxy = null;
-
-        try
-        {
-            // *** SETUP *** \\
-            BasicHttpsBinding binding = new BasicHttpsBinding(BasicHttpsSecurityMode.Transport);
-            binding.Security.Transport.ClientCredentialType = HttpClientCredentialType.None;
-            endpointAddress = new EndpointAddress(new Uri(Endpoints.Https_SecModeTrans_ClientCredTypeNone_ServerCertValModePeerTrust_Address));
-
-            factory = new ChannelFactory<IWcfService>(binding, endpointAddress);
-            factory.Credentials.ServiceCertificate.SslCertificateAuthentication = new X509ServiceCertificateAuthentication();
-            factory.Credentials.ServiceCertificate.SslCertificateAuthentication.CertificateValidationMode = X509CertificateValidationMode.PeerOrChainTrust;
 
             serviceProxy = factory.CreateChannel();
